Split relational assertions into completeness, uniqueness, integrity
