+++ src/moneybin/validation/assertions/__init__.py
@@ -5,6 +5,7 @@ from moneybin.validation.assertions.business import (
     assert_date_continuity,
     assert_sign_convention,
 )
+from moneybin.validation.assertions.completeness import assert_no_nulls
 from moneybin.validation.assertions.distributional import (
     assert_distribution_within_bounds,
     assert_unique_value_count,
@@ -15,9 +16,7 @@ from moneybin.validation.assertions.infrastructure import (
     assert_no_unencrypted_db_files,
     assert_sqlmesh_catalog_matches,
 )
-from moneybin.validation.assertions.relational import (
-    assert_no_duplicates,
-    assert_no_nulls,
+from moneybin.validation.assertions.integrity import (
     assert_no_orphans,
     assert_valid_foreign_keys,
 )
@@ -27,6 +26,7 @@ from moneybin.validation.assertions.schema import (
     assert_row_count_delta,
     assert_row_count_exact,
 )
+from moneybin.validation.assertions.uniqueness import assert_no_duplicates
 
 __all__ = [
     "assert_balanced_transfers",
